+++ R/Ellipse.R
@@ -758,3 +758,28 @@ LownerJohnEllipse <- function(pts){
   EllipseFromCenterAndMatrix(-c(chol2inv(chol(B)) %*% y[c(4L,5L)]),
                              tcrossprod(B))
 }
+
+#' @title Smallest ellipse that passes through three boundary points
+#' @description Returns the smallest area ellipse which passes through
+#'   three given boundary points.
+#'
+#' @param P1,P2,P3 three non-collinear points
+#'
+#' @return An \code{Ellipse} object.
+#' @export
+#'
+#' @examples
+#' P1 <- c(-1,0); P2 <- c(0, 2); P3 <- c(3,0)
+#' ell <- EllipseFromThreeBoundaryPoints(P1, P2, P3)
+#' ell$includes(P1); ell$includes(P2); ell$includes(P3)
+EllipseFromThreeBoundaryPoints <- function(P1, P2, P3){
+  if(.collinear(P1, P2, P3)){
+    stop("The three points are collinear.")
+  }
+  points <- rbind(P1, P2, P3)
+  means <- colMeans(points)
+  cpoints <- sweep(points, 2L, means)
+  S <- 1.5 * solve(crossprod(cpoints))
+  EllipseFromCenterAndMatrix(means, S)
+}
+